requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy>=1.1.0
cryptography==41.0.5
cycler==0.12.1
dataclasses-json==0.6.2
distro==1.8.0
dnspython==2.4.2
effdet==0.4.1
emoji==2.8.0
exceptiongroup==1.1.3
faiss-cpu==1.7.4
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.45.0
frozenlist==1.4.0
fsspec==2023.10.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.4
iopath==0.1.10
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.0.339
langdetect==1.0.9
langsmith==0.0.66
layoutparser==0.3.4
loguru==0.7.2
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib>=3.0,<4.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx>=2.0,<3.0
nltk==3.8.1
numpy>=1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==1.3.3
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.3
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.3
Pillow==10.1.0
pinecone-client==2.2.4
portalocker==2.8.2
protobuf==4.25.1
pycocotools==2.0.7
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
pyparsing==3.1.1
pypdf==3.17.1
pypdfium2==4.24.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-iso639==2023.6.15
python-magic==0.4.27
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.5.2
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.1
timm==0.9.11
tokenizers==0.15.0
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.35.2
triton==2.1.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
unstructured==0.11.0
unstructured-inference==0.7.15
unstructured.pytesseract==0.3.12
urllib3==2.1.0
wrapt==1.16.0
yarl==1.9.3
